64 bit align instructions
Summary:
Some instructions store `Symbol`. `Symbol` has a field of type `u64`. So instructions need to be 64-bit aligned even on 32 bit.

Alternatively we can implement some machinery of keeping headers aligned to 32 bit, but align 64 bit instruction content for instruction which need it, but that is harder to implement, and slower at runtime.

Also add a bunch of assertions when helped to debug this code when running on 32-bit.

Reviewed By: ndmitchell

Differential Revision: D46430383

fbshipit-source-id: 2eb84b95fc2a676a50c2019f151708bda088e65a