Use "correct" frame id for "Root" frame
Summary:
For evaluation we will reject any evaluation in a non-zero frame.
There are situations where we only have the Root frame, and so hardcoding it to
a large frame id breaks debugger evaluation in those situations. Instead, we
can just compute the appropriate id based on length of the frames vec.

Reviewed By: wendy728

Differential Revision: D46041445

fbshipit-source-id: 9adaa719d93eb1301cc5bf536bff3d8f217e8e40