Slightly tidy up doc to type conversion
Summary: I ran a script to eyeball the differences in primitives between the doc oracle and the one I wrote by hand. This was one of the differences. I'm not sure it ever matters in practice, but for `min` it results in a slightly different type for the key parameter.

Reviewed By: krallin

Differential Revision: D46677181

fbshipit-source-id: 6efa9b4cea9eaae90017305aa7bd5701963a3310