Make Value.to_int do its own conversion
Summary: The `to_int` stuff is pretty dubious - convert to an int ignoring that some things are bool. It's mostly used for original Python functions where the stdlib did the wrong thing, and we retain compatibility. Instead of putting it on StarlarkValue we should just had code the few instances, and try and eliminate it where we can.

Reviewed By: wendy728

Differential Revision: D46685674

fbshipit-source-id: 6aa8779a98b0e2328ce02ba878acc5a887c22160