Return an error instead of panic on incorrect invariants in typechecker
Summary:
Plugging typechecker into the compiler in dangerous: small bug can result in error.

Panicking even on internal errors is not optimal because if buckd crashes, not much information is available even what file causes the error.

I propose using more explicit errors for state validation.

In this case, it is not strictly necessary, but this illustrates the approach, which I plan to use more in typechecker.

Note, explicit type `InternalError` is used to not mistake it for regular typechecker error.

Reviewed By: ndmitchell

Differential Revision: D46847976

fbshipit-source-id: f246fc7160aafdd4b5623886f996227662bca97c