#[starlark(as_type = ...)] attribute
Summary:
Can be used like this:

```
    #[starlark(as_type = StarlarkStr, speculative_exec_safe)]
    fn str<'v>(
        #[starlark(require = pos)] a: Value<'v>,
        eval: &mut Evaluator<'v, '_>,
    ) -> anyhow::Result<StringValue<'v>> {
```

`as_type` should point to the implementation of `StarlarkValue`.

Currently `StarlarkValue` is only used to obtain a type name string, but eventually it should be used to fetch the implementation of type: when typechecking code:

```
def foo(r: str): r + "x"
```

(note no `.type`), typechecker should fetch `str` value from global, and fetch the reference to `StarlarkValue` implementation of string (`StarlarkStr`), which can be used to typecheck the function without help of oracles.

Reviewed By: ndmitchell

Differential Revision: D47001469

fbshipit-source-id: 4025138487e28e16c4ede25b094f1ab82706491d