Prohibit d[1, 2]
Summary:
In Python, expression:

```
d[1, 2]
```

is equivalent to

```
d[(1, 2)]
```

And starlark-rust had it too before this diff (TIL and there are no tests for it).

[Starlark spec](https://github.com/bazelbuild/starlark/blob/master/spec.md#index-expressions) prohibits it and I agree with it.

Let's prohibit it too.

Hope nothing will break.

**Context**

Implementing type expression `dict[str, int]`. We could reuse current behavior (removed in this diff), but maybe better only allow this syntax for type expressions.

The plan is to add a method:

```
trait StarlarkValue {
  // existing
  fn at(&self, index: Value, heap: &Heap) -> Result<Value> { ... }

  fn at2(&self, index0: Value, index1: Value, heap: &Heap, private: Private) -> Result<Value> { ... }
}
```

and implement `at2` function only for `dict` value.

So `dict[(str, int)]` will be an error unlike Python where it is identical to `dict[str, int]`.

Reviewed By: ndmitchell

Differential Revision: D47085133

fbshipit-source-id: b07cc83467affe60c26e2fcd8a167d8fa71a6385