remove links to archived google/skylark repo from docs
Summary:
I'm not sure what value these links were providing... the docs they link to say the same thing. they are also redundant because the function name is already a header in the rendered docs.

since the google/skylark repo they link to is archived, it makes it look stale.

Reviewed By: bolinfest

Differential Revision: D47632443

fbshipit-source-id: b64b04d61e4142ba887fb532a7f63112eda2a998