Lex comments
Summary:
We need to implement proper tooling with our starlark. For example, linters with autofix, or codemod tools.

These tools depend on code formatting.

It is not clear yet, how to implement code formatting, but to do code formatting we need to preserve comments while parsing.

This diff adds comment token to the lexer. Which is dropped later when parsing.

Don't know yet how to implement code formatting. I see two options:
* either add comments to each AST node (as [buildifier does it](https://github.com/bazelbuild/buildtools/blob/b31f2c13c407575100d4614bcc9a3c60be07cc1c/build/syntax.go#L735-L736)). This is safer, but makes AST and verbose (in parsing, in compilation)
* or store comments separately and while printing AST, read comments. This does not allow manipulating comments programmatically.

Reviewed By: krallin

Differential Revision: D47701375

fbshipit-source-id: 1bd8308629ff256c5e3584e437e13179621e5b05