Change record type comparison with identity
Summary:
If we have two record definitions, in different `bzl` files:

```
Rec = record()
```

We consider them equal (and instances of these records equal if fields are equal) **if** record type names are equal.

This does not look right: we should either do structural equality only, or consider identical record types defined in different bzl files not equal.

The latter approach seems to be more correct: `record` functions defines a type, so two calls should define two types, and instances of these types should not be considered equal because types are not equal.

This change is needed not to just fix this issue, but also for typing: **for typing we want to use an instance of a record type as a type descriptor** (for both static and runtime typechecker).

This structural equality behavior was added in D29303210 (which does not explain much), and this diff basically reverts it.

Maybe something may break when a rule is defined in one cell, but used in another cell? I don't see in code how.

I could not find any scenarios that breaks after this change.

Perhaps it is because of this change D34380414 in buck2 prelude handling.

Reviewed By: krallin

Differential Revision: D47115144

fbshipit-source-id: 1da9b96e046d87353eef496d3599099f207fc908