Store CanonicalTypeId in AValueVTable
Summary:
`StarlarkTypeId` (D48090538) is canonical id of the type, e.g. `Dependency<Value>` and `Dependency<FrozenValue>` should share the same `StarlarkTypeId`.

The idea is this: both `Value` and `TyStarlarkValue` store `StarlarkTypeId`, so runtime isinstance checking for `TyStarlarkValue` can be done by comparing that `StarlarkTypeId`.

This is later, but for now use `StarlarkTypeId` for struct instance check. This saves like two instructions per isinstance check, but illustrates the idea.

Reviewed By: ndmitchell

Differential Revision: D48090857

fbshipit-source-id: db8e48ba5683b90cf5ee8d731a62a9d5cd054e99