Evaluate ...
Summary:
We don't have a way to describe a type tuple of unknown length with given element types. In Python they use `tuple[str, ...]`. Probably we want the same.

This type is the type of parameter in:

```
def f(*x: str): pass
```

(Technically we don't need to express this type in starlark, but we probably should be able to express any type in starlark.)

This diff adds `...` expression when types are enabled in dialect.

Reviewed By: ndmitchell

Differential Revision: D48319025

fbshipit-source-id: 5741f2359989c1b163d2aa212af9341fda64f16e