Only typecheck top-level defs
Summary:
Instead of typechecking the whole module at once, iterate only top-level statements and typecheck defs.

The reasons are these:
* Top-level statements are evaluated anyway, not much point in typechecking them
* This is how typechecking should be done:
  * evaluate/typecheck top-level assignments
  * then evaluate defs
  * This diff moves to that direction, except for now there's no typechecking of top-level assignments
* It is faster, because

Currently typechecker is implemented by iterating up to 100 times all statements until types converge (I don't understand it, ask Neil). So we do the number of iteration equal to the number of iteration needed to typecheck the worst function. With each function typechecked independently, average number of iteration is smaller.

It does not matter much for lightweight typechecker, but it is critical for compiler typechecker (following diff D48752027), and this diff is done to make both typecheckers uniform.

Reviewed By: ianlevesque

Differential Revision: D48752028

fbshipit-source-id: 7909e192aff1bbf0afc3a14a8444073a7690ad0c