Prohibit strlit in type expr
Summary:
Prohibit string literals in type expressions unless `starlark-rust: allow_string_literals_in_type_expr` annotation is present.

Previous diffs in this stack add this annotations to every file in fbsource where the annotation is present.

This is to:
* track and migrate string literals in type expressions to proper types
* prevent regressions

Note this does not prohibit string constants in type expressions, but we need to start somewhere.

Reviewed By: cjhopman

Differential Revision: D48916740

fbshipit-source-id: 178a672af4f5454fd42525deb9d6d39b32361e1e