Require only StarlarkTypeRepr in StarlarkValueAsType
Summary:
`StarlarkValueAsType` utility is used to declare globals which are type symbols.

With this diff, it is possible to use any `StarlarkTypeRepr` to declare a symbol, not just `StarlarkValue`.

The diff is large because it changes internals. Previously `StarlarkValueAsType<T>` was a `StarlarkValue` itself.

Now `StarlarkValueAsType<T>` only implements `AllocFrozenValue` and the actual `StarlarkValue` implementation lives in a separate struct `StarlarkValueAsTypeStarlarkValue`.

Reviewed By: ianlevesque

Differential Revision: D48892725

fbshipit-source-id: 2a1f4843b15e10e0c9c6692a081faabde625b295