Make any TyName intersect with any TyCustom
Summary:
Following diff D48893103 implements `Provider` type as `TyUser` (which is `TyCustom`), but does not yet convert concrete provider types to `TyUser`. So we need to intersect `TyUser` with `TyName`.

This makes typechecker a bit weaker than before, but that's fine because the plan is to get rid of `TyName`.

Reviewed By: ianlevesque

Differential Revision: D48893554

fbshipit-source-id: 976dd58685cf7f11e648b812d73cf07631920ae9