Fix the issue with TypeId alignment on new Rust Apple Silicon
Summary:
* `TypeId` is `u128` [in Rust 1.72](https://github.com/rust-lang/rust/pull/109953/)
* `u128` is 16 bytes aligned on Apple Silicon
* we require 8 byte alignment for `StarlarkValue`

Maybe we should fix heap allocations: insert padding to allow any alignment.

But practically we don't need it and should not do it: generated machine code is identical for aligned and 8-byte aligned `u128`: [compiler explorer](https://rust.godbolt.org/z/96KsnjW77) (note it is identical if we load it from memory not pass by value, but this is what we do).

So for now just force 8 byte alignment on `TypeId`.

Fixes issue https://github.com/facebook/buck2/issues/408

Reviewed By: ndmitchell, shayne-fletcher

Differential Revision: D49019964

fbshipit-source-id: 62c8d7db83e9e4bdbe58fe7a62d5daae41996e61