Make AST private
Summary:
Make fields of `AstModule` private.

Expose them with `AstModuleFields` trait, which is not visible outside of `fbcode//buck2/starlark/...`.

AST is already expensive to maintain, if it is public, it will be even harder.

If something is needed from AST (like lints), it should be exposed as public API of `starlark` crate (not `starlark_syntax`).

Reviewed By: ndmitchell

Differential Revision: D49128233

fbshipit-source-id: f344e49be0438fa820cf24a1c886624c0b9f0b4a