add call_stack function
Summary:
For complex rules this is very useful to provide extra context to engineers for
debugging failing builds.

More specifically, I want to record a stacktrace at parse time from a macro
that I can then display to a user from a Rust program that is used in
`ctx.actions.run` to provide additional context of where exactly an invalid
definition was generated, otherwise it can be very hard to find.

Reviewed By: wendy728

Differential Revision: D50203670

fbshipit-source-id: 326d3b53ebc36c101abe955e976a04d92e9c153d