errors: `starlark::Error` in `starlark_bin`
Summary: The goal here is not to switch the entire `starlark_bin` crate over to `starlark::Error` (`anyhow::Error` is probably fine for most of it). This diff just focuses on switching a couple places which produce errors that are later consumed by `EvalMessage`s. Ensuring that we have a `starlark::Error` from which to create an `EvalMessage` is important

Reviewed By: stepancheg

Differential Revision: D51573517

fbshipit-source-id: d926acb3ef7096c3a3a80b66f9e18b132d2ca2df