errors: `ErrorKind::Value`
Summary:
Let's start having some non-trivial error kinds.

We need to use the type erasure here because `ErrorKind` is defined in `starlark_syntax`. If we removed the `unsupported_with` method from `ValueError` we could move `ValueError` to `starlark_syntax`, but I don't know if that's an improvement. Happy to take feedback on that

Reviewed By: stepancheg

Differential Revision: D51735680

fbshipit-source-id: 105cfa47104aa2be97dd2d3472e4e836b12e8d0a