Fix "go to label" in bazel LSP
When executing "go to definition" on a label, it would previously go to the correct file but not to the definition of that target. The reason was because we were searching for a function call where the name is the whole label. Now it parses the label to extract the name and searches using that.

This doesn't work in a few cases, for example "//foo" and "@foo", where the name is implicit. However, I think to solve this we would need to implement a label parser. If one already exists, I could not find it. I imagine a lot of the code in `resolve_folder` could be simplified with the use of a label parser too.