Bazel: Fix relative go to definition in external repositories
For relative loads from external repositories, the resolve root was calcualated incorrectly. Instead of _removing_ the path of the file within the external repository, it only kept this part. For example if the current file was `$(bazel info output_base)/external/foo/bar/BUILD`, it would compute the resolve root as `bar/BUILD`, whereas it needs to be `$(bazel info output_base)/external/foo`.