Fix "go to label" in bazel LSP (#100)
Summary:
When executing "go to definition" on a label, it would previously go to the correct file but not to the definition of that target. This was because we were searching for a function call where the name is the whole label. Now it parses the label, using the label parser introduced in https://github.com/facebookexperimental/starlark-rust/issues/101, to extract the name and searches using that.

Pull Request resolved: https://github.com/facebookexperimental/starlark-rust/pull/100

Reviewed By: wendy728

Differential Revision: D52954185

Pulled By: ndmitchell

fbshipit-source-id: 937da257b7441d8abe313015ca1c8241c7967a10