add support for lint suppressions
Summary:
In order to make starlark a "trusted linter", i.e. a linter for which we can enforce the codebase to be 100% clean, we need a facility to suppress warnings inline in the code.

This will take take the shape of

```
def foo() -> str: # starlark-lint-disable missing-return
  pass

# starlark-lint-disable missing-return
def bar():
  pass

def baz() -> str
  pass # starlark-lint-disable missing-return
```

Multiple 'short_names' can be separated on a single line.

```
# starlark-lint-disable missing-return, missing-return-expression
def foo() -> str:
  pass
```

Multiple supression lines will accumulate if put in a block of comments

```
# We need to do this
# starlark-lint-disable missing-return, unreachable
# and this, sorry
# starlark-lint-disable missing-return-expression
# another comment
def foo() -> str:
  pass
```

Reviewed By: ndmitchell

Differential Revision: D54072440

fbshipit-source-id: 0722afba1a2b6e76c2d3f0a31edc97ee0a65dc37