Add back the third parameter to Evaluator
Summary:
This actually existed before and was removed > 3 years ago with the reasoning:

"It didn't really let you do anything more, since the value was &'a
AnyLifetime<'e>, so 'e couldn't outlive 'a. At the same time, you could pick
the 'a freely, so you didn't ever need them to be different."

That's only true though when the type we put there is covariant in 'e. If it's
invariant (like if it holds a `RefCell<&'e mut X>`) then it cannot fit easily there without
a separate lifetime.

Reviewed By: JakobDegen

Differential Revision: D54214064

fbshipit-source-id: 162e12e007f794c0ea3aaaaf2af4f281e8ca396d