feat: add eager flag and workspace symbols feature
Allow passing in an `--eager` flag to the LSP to make it eagerly load
all relevant files in the workspace. When this flag is provided, files
are not removed from the `last_valid_parse` cache when they are closed.

In addition, this flag enables the LSP to provide workspace symbol
information.