Do not require Display for T in StarlarkAny<T>
Summary:
Types put into `StarlarkAny<T>` usually do not appear as normal values to starlark code, but hidden in other values. So `Display` is not very helpful there.

(`Display` is needed for starlark vtable, and called for `repr(x)`).

There's some overengineering here:

https://www.internalfb.com/code/fbsource/[1c722b9f847c1bdf1a70a43fa8748c32518d689e]/fbcode/buck2/starlark-rust/starlark/src/values/layout/heap/heap_type.rs?lines=423%2C433

which I would prefer remove.

Because I want to implement different version of `StarlarkAny`: with `Trace` and `Freeze`, and avoid replicating this machinery.

In the bright future when we have specialization in Rust, we can inherit `Display` if it is implemented and use `Debug` otherwise, but that probably won't happen in our lifetimes.

Reviewed By: JakobDegen

Differential Revision: D56033101

fbshipit-source-id: 3b40fb2ae2b95bb70bce71fffd1e6e8aaeef64a0