Do not hang in debugger tests if evaluator fails
Summary:
When evaluator exits with error, closure is dropped, reset `breakpoints_hit` number to 999999, and fail any following operation accessing this `breakpoint_hit`.

Context: I patched the interpreter, it started to fail, but tests started to hang instead of failing, because operations like

```
controller.wait_for_eval_stopped(2, TIMEOUT);
```

never finished because the interpreter already quit.

This is not perfect fix, out of 5 tests, one still hanged with this changes and my breaking the interpreter. But better than nothing.

Reviewed By: VladimirMakaev

Differential Revision: D58389188

fbshipit-source-id: c0ce78e1afd2faa55d30000a59168bcee21af47a