Fix unwrap panic in bazel-lsp for Label (#121)
Summary:
Fix for https://github.com/cameron-martin/bazel-lsp/issues/41 - Panic when rendering Label documentation

Looks like documentation may contain no lines (in case of bazel-lsp at least), and this quick fix makes it work. Not sure whether it's the best solution, or the assumption that docs should contain non-empty text should be enforced somewhere, but it is not at this point so making it not panic seems like good solution.

Pull Request resolved: https://github.com/facebook/starlark-rust/pull/121

Reviewed By: perehonchuk

Differential Revision: D59208773

Pulled By: ndmitchell

fbshipit-source-id: 00708614a9e4fb53e1eaf4f41bedcae1fd4176f9