Remove unused capacity support from profiler
Summary:
Unused capacity attribute for frozen heaps was useful before unused capacity became reused in the following allocation (D40738710).

Now it is computed incorrectly (it is always zero; technically it is not zero when chunks are not reused, but we cannot know that where profile is generated).

Unused capacity is slightly useful for "allocated" profiler, but not much (because with "allocated" we measure loading, and release memory immediately after). And it is not free to maintain, so I propose to just drop it to have less code to maintain.

Context:

Before this diff recursive profile analysis looked like this:

 {F1741730275}

Reviewed By: JakobDegen

Differential Revision: D59345258

fbshipit-source-id: f3af4b3af88c9e8eee6044d736713953b33abeee